言語設定タイルをコンポーネント化
