+++ lib/feature/github/view/part/drawer_widget.dart
@@ -35,43 +35,7 @@ class DrawerWidget extends ConsumerWidget {
             ),
           ),
           // 言語設定ボタン
-          ExpansionTile(
-            title: const Text('言語設定'),
-            children: [
-              RadioListTile<Locale>(
-                title: const Text('日本語'),
-                value: const Locale('ja'),
-                groupValue: currentLocale,
-                onChanged: (Locale? value) {
-                  _setLocale(value, ref);
-                },
-              ),
-              RadioListTile<Locale>(
-                title: const Text('English'),
-                value: const Locale('en'),
-                groupValue: currentLocale,
-                onChanged: (Locale? value) {
-                  _setLocale(value, ref);
-                },
-              ),
-              RadioListTile<Locale>(
-                title: const Text('한국어'),
-                value: const Locale('ko'),
-                groupValue: currentLocale,
-                onChanged: (Locale? value) {
-                  _setLocale(value, ref);
-                },
-              ),
-              RadioListTile<Locale>(
-                title: const Text('中文'),
-                value: const Locale('zh'),
-                groupValue: currentLocale,
-                onChanged: (Locale? value) {
-                  _setLocale(value, ref);
-                },
-              ),
-            ],
-          ),
+          LanguageToggleTile(currentLocale: currentLocale),
           ListTile(
             title: const Text('ログアウト'),
             onTap: () async {
@@ -87,12 +51,6 @@ class DrawerWidget extends ConsumerWidget {
         ],
       ),
     );
-  }
-
-  void _setLocale(Locale? value, WidgetRef ref) {
-    if (value != null) {
-      ref.read(localeNotifierProvider.notifier).setLocale(value);
-    }
   }
 
   void _logout(BuildContext context, WidgetRef ref) {
